requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-versatileimagefield==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
idna==2.10
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.2.0
pycparser==2.20
pygraphviz==1.7
PyJWT==2.1.0
python-magic==0.4.22
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
